refactor(matrix): move verification code into submodule
Owning abstraction: Matrix verification request ingestion, action routing, SAS projection, pruning, and rate-cap admission live in channels::matrix::verification.

Invariant: verification DTO wire shapes, raw protocol-flow-id SDK lookups, trust gates, timeout refreshes, and WS broadcasts preserve their existing behavior after the move.

Validation: /Users/zera/build/carapace/scripts/cargo-serial fmt --check; /Users/zera/build/carapace/scripts/cargo-serial check --all-targets; /Users/zera/build/carapace/scripts/cargo-serial nextest run verification.
